requirements.txt
Django
python-decouple
python-dotenv
openai
boto3
botocore
django-storages
pygame
opencv-python
pydub
pytest
pylint
sqlfluff
langchain
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.3.0
reportlab
python-docx
openpyxl
python-pptx
mcp>=1.2.0
requests
playwright
pydantic
duffel-api
chromadb